Updated database-connecting PHP files to depend on mysql_connect.php file
