+++ ui/challenge.php
@@ -1,17 +1,6 @@
 <?php
-	function dbConnect()
-	{
-		$dbConn = mysql_connect('', '', '');
-		if ( ! $dbConn )
-		{
-			die("Unable to connect to database: " . mysql_error());
-		}
-		$db_selected = mysql_select_db('', $dbConn);
-		if ( ! $db_selected )
-		{
-			die ("Unable to select database: " . mysql_error());
-		}
-	}
+	require('mysql_connect.php');
+	dbConnect();
 	
 	session_start();
 ?>
